[BE] feat(world): Session 엔티티 session_id 타입 변경 (S14P21C103-191) - game_session 테이블에 맞춰 long -> int 로 변경

+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/domain/world/housing/GameContractReviewRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface GameContractReviewRepository extends JpaRepository<GameContractReview, Long> {
 
     Optional<GameContractReview> findTopByGameSessionIdAndPropertyIdOrderByReviewedAtDesc(
-        Long gameSessionId,
+        Integer gameSessionId,
         Long propertyId
     );
 }
